backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.0
pydantic==1.10.8
bcrypt==4.0.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
passlib[bcrypt]
